internal/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
PyYAML>=6.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
chromadb>=0.4.0
httpx>=0.25.0
pdfplumber>=0.10.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
